Fixed File Name Issue
Changes:
* Fixed an issue on case-insensitive operating systems with the GameObject header files.
